refactor(spa): extract form field components to hosts/form-fields (#138, #185)
